Fixing issue with reporting poll comments

+++ protected/modules/bigbrother/views/admin/view.php
@@ -14,22 +14,34 @@ use humhub\modules\polls\models\Poll;
             <strong>Reported Content</strong>:<br/>
             <div class="well">
                 <?php
-                    $content = Content::findOne(['id'=>$report->content_id]);
-                    switch(get_class($content->getModel())) {
-                        case Poll::class:
-                            $poll = json_decode($report->message);
-                            echo "Question: " . $poll->question . "<br/>";
-                            echo "Description: " . $poll->description . "<br/>";
-                            echo "Answers:<br/><ul>";
-                            foreach ($poll->answers as $answer)
-                            {
-                                echo "<li>".$answer."</li>";
-                            }
-                            echo "</ul>";
-                            break;
-                        default:
-                            echo $report->message;
-                            break;
+                    if($report->comment_id) {
+                        echo $report->message;
+                    } else {
+                        $content = Content::findOne(['id'=>$report->content_id]);
+                        switch(get_class($content->getModel())) {
+                            case Poll::class:
+                                $poll = json_decode($report->message, true);
+                                if(is_array($poll)) {
+                                    echo array_key_exists('question', $poll) ? "Question: " . $poll['question'] . "<br/>" : '';
+                                    echo array_key_exists('description', $poll) ? "Description: " . $poll['description'] . "<br/>" : '';
+                                    if(array_key_exists('answers', $poll)) {
+                                        echo "Answers:<br/><ul>";
+                                        foreach ($poll['answers'] as $answer)
+                                        {
+                                            echo "<li>".$answer."</li>";
+                                        }
+                                        echo "</ul>";
+                                    }
+                                } else {
+                                    echo "<pre>";
+                                    var_dump($poll, $report);
+                                    echo "</pre>";
+                                }
+                                break;
+                            default:
+                                echo $report->message;
+                                break;
+                        }
                     }
                 ?>
             </div>
